requirements.txt
GitPython
audioop-lts
py-cord
python-dotenv
